refactor: move update.js mapPin.svg to common
refactor INTERN_DASHBOARD_URL

#160
